Added proper time to messages

+++ Server/DataStructures/PlayerMessage.cs
@@ -4,7 +4,7 @@ namespace MinimalGameServer.DataStructures
 {
     public class PlayerMessage
     {
-        public float TimeSent { get; set; }
+        public long TimeSent { get; set; }
         public string Content { get; set; }
         public string Author { get; set; }
         public PlayerMessage() 
@@ -13,14 +13,14 @@ namespace MinimalGameServer.DataStructures
             Content = "No content";
             Author = "Unknown";
         }
-        public PlayerMessage(Player player, float timeSent, string content)
+        public PlayerMessage(Player player, long timeSent, string content)
         {
             Author = player.Name;
             TimeSent = timeSent;
             Content = content;
         }
         [JsonConstructor]
-        public PlayerMessage(string Author, float TimeSent, string Content) 
+        public PlayerMessage(string Author, long TimeSent, string Content) 
         {
             this.Author = Author;
             this.TimeSent = TimeSent;
